% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tera-package.R
\docType{package}
\name{tera-package}
\alias{tera}
\alias{tera-package}
\title{tera: Tera Templating Engine}
\description{
The 'tera' package uses 'extendr' to provide access to Vincent Prouillet's 'Tera' templating engine in Rust. Users mainly interact with a Tera R6 object, which serves as a template library with encapsulated methods for rendering templates with a given context. Template syntax supports additional logic, including built-in filters, tests, and functions, as well as loops, conditions, and inheritance. Documentation for Tera's templating syntax can be found at \url{https://keats.github.io/tera/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbvernon/tera-r}
  \item \url{https://kbvernon.github.io/tera-r/}
}

}
\author{
\strong{Maintainer}: Kenneth Blake Vernon \email{kenneth.b.vernon@gmail.com} (\href{https://orcid.org/0000-0003-0098-5092}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kenneth Blake Vernon \email{kenneth.b.vernon@gmail.com} (\href{https://orcid.org/0000-0003-0098-5092}{ORCID}) [copyright holder]
}

}
\keyword{internal}
