% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-template.R
\name{render-one-off}
\alias{render-one-off}
\alias{render_template}
\alias{render_string}
\title{One-Off Template Rendering}
\usage{
render_template(path, ..., outfile = NULL, autoescape = TRUE)

render_string(string, ..., outfile = NULL, autoescape = TRUE)
}
\arguments{
\item{path}{character scalar, path to a template file}

\item{...}{specify context as key-value pairs where key is the template
variable and value is the data to inject.}

\item{outfile}{character scalar, the path to file where template is to
be rendered. If \code{NULL} (the default), it will render the template file to a
string in the current R session.}

\item{autoescape}{boolean scalar, whether to autoescape HTML (default is
\code{TRUE}).}

\item{string}{character scalar, the template string to render.}
}
\value{
`outfile`` (invisibly)
}
\description{
For rendering a single template file or string, it may be
preferable to use these one-off rendering options.
}
\details{
Requires a path to a template file, not a template string.
}
\examples{
tmp <- system.file("templates/hello.html", package = "tera")

# render to string
render_template(
  tmp,
  x = "world",
  y = "tera"
)

render_string(
  "<p>Hello {{ x }}. This is {{ y }}.</p>",
  x = "world",
  y = "tera"
)

# render to file
outfile <- file.path(tempdir(), "hello-rendered.html")
render_template(
  tmp,
  x = "world",
  y = "tera",
  outfile = outfile
)

readLines(outfile, warn = FALSE)
}
